# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FurniCore ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created — same shape as login response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Auth tokens + user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshBody"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /employees:
    get:
      operationId: listEmployees
      tags: [hr]
      summary: List employees
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [hr]
      summary: Create employee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [hr]
      summary: Get employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [hr]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [hr]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/attendance:
    get:
      operationId: getEmployeeAttendance
      tags: [hr]
      summary: Get attendance for employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance:
    post:
      operationId: recordAttendance
      tags: [hr]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceBody"
      responses:
        "201":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # PAYROLL
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [accounting]
      summary: List transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [accounting]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [accounting]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # NOTIFICATIONS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    RefreshBody:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - refreshToken
        - user

    # USERS
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        hireDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - department
        - position
        - baseSalary
        - hireDate
        - isActive
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        hireDate:
          type: string
        userId:
          type: integer
      required:
        - name
        - email
        - department
        - position
        - baseSalary
        - hireDate
    UpdateEmployeeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        position:
          type: string
        baseSalary:
          type: number
        isActive:
          type: boolean
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        date:
          type: string
        status:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        hoursWorked:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - employeeId
        - employeeName
        - date
        - status
        - createdAt
    CreateAttendanceBody:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        status:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        notes:
          type: string
      required:
        - employeeId
        - date
        - status

    # PAYROLL
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        reference:
          type: ["string", "null"]
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        status:
          type: string
        transactionDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - category
        - amount
        - description
        - status
        - transactionDate
        - createdAt
        - updatedAt
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        reference:
          type: string
        supplierId:
          type: integer
        transactionDate:
          type: string
      required:
        - type
        - category
        - amount
        - description
        - transactionDate

    # NOTIFICATIONS
...